Oops

+++ include/chemskr/Lexer.h
@@ -13,7 +13,7 @@ using CHEMSKRSTYPE = Chemskr::ASTNode *;
 #include "formula.h"
 #endif
 
-using yy_size_t = size_t;
+using yy_size_t = int;
 
 using leng_t = int;
